Fix: Add TestCharge class in MavenTests/main.cpp
MavenTests/main.cpp runs the tests in the application. If a test
class is not present in the MavenTests/main.cpp, the test won't
run and hence the XML to check pass/fail won't be created.

Adding TestCharge class into MavenTests/main.cpp creates a file
named testCharge.xml to check passing/failing of tests