UI: Update qsettings & mavenParameters from UI #151
QT Settings are updated when something is changed in the UI of
"Options" Dialog. QT Setting may or maynot update the maven
Parameters class object but it is recommended to do so to follow
MVC Architecture. If maven parameters class object has to be
updated then updated QT Settings are being used.

Used four variables:
1. distXWeight - weight of distX
2. distYWeight - weight of distY
3. overlapWeight - weight of overlap
4. useOverlap - option to select/de-select overlap for group scoring

QT Settings & mavenParameters are updated with these variables.

Issue: #151