Ftr: Add quantile filter for all peak filter parameters #267
In peak filtering, a group is accepted if the group maximum for every
parameter is above the user input thresholds. This is being modified to
ensure that a group is accepted only if a minimum percentage of its peaks
lie above the thresholds; else the group is filtered out.

Quantile filters for Peak Intensity and Peak Quality already exist. They
have now been added for two more parameters:
- Signal/Baseline ratio
- Signal/Blank ratio
The remaining parameters do not require this addition

Issue: #267