Fix: Add a flag for deltaRt weight to replace matchRt flag in group rank calculation #217
Match RT checkbox in the Peak Detection Dialog was used to decide if Delta RT weight should be considered for Group Rank calculation. "Consider deltaRT weight" in the Group Rank tab of Options dialog is now used in place of match RT for group rank calculation. Match Rt continues to be used for other purposes.

Issue: #217