Ftr: Assign MetagroupId for each Isotopic set of groups #266
MetagroupId for all the groups are zero all over the code.
So, we are aassigning them with some values starting from 1 as follows.

For each different parentGroup(for all the isotopic children groups),
get a new MetagroupId by incrementing. Thus, all the children of
that parent will also get the same MetagroupId.