Fix: No polarity checking when charge is zero in DB #260
In function "findSpeciesByPrecursor", Compounds are rejected if
polarity (parameter) is not same as charge of the Compound. When
polarity is not present inside Compound Database file, Compound
charge is set to zero. But when compound charge is zero, user
Polarity is set to some default value in function getSrmSlices.

No comparing user polarity and compound charge when compound charge
is zero

Issue: #260