Ftr: set filterSignalBaselineDiff in EIC #153
In function getPeakPositions, peaks can also be filtered
on the basis of signal to baseline difference. A function in
EIC class called setFilterSignalBaselineDiff() can be used
to set the minimum value of signal to baseline difference. If
minSignalBaselineDifference is below this set value, peak will
be rejected and not marked as peak.

set filter for signal to baseline difference using function
setFilterSignalBaselineDiff() before getPeakPositions() function.
Made the required changes in the whole code to incorporate this change.

Issue: #153