Test: Modify test case for addGroups #314
addGroups() in csvreports adds all group information (grup formula, compound id, maxQuality etc) to the group report

testaddGroup was comparing each of these parameters with standard values. This test would fail if there was an error in grouping even though addGroups is working as expected.

This commit changes the test condition to only check the number of columns and the group id.

Issue: #314