Doc: Give a good commit message subject line #324
This commit message example is a good example of its own. Its  subject
line has 50 or less than 50 characters and body lines have 72 or  less
than 72 characters.This paragraph  of this message  is explaining what
changes I'm making, next paragraph is going to  explain why I'm making
this change and next of next one is going to  explain  how I'm  making
this change.

Proper commit message is neccessary to understand the changes someone is
making and  impact  of these changes  to the codebase. It  helps fellow
developers to understand the codebase and to save their time for faster
delivery of product. It also evaluates colaboration skill of a developer.

* First I have provided instructions to write good commit message.
* Next I have given an example.
* In this section I'm using bullet points,it could be "-" or "*" .
* Bullet points are not neccessary. It can also be simple paragraph.
* Last line have tag for issue this commit is going so resolve.
* There is no other issue related to this issue, that's why there is
  no "See also: #IssueNo, #IssueNo ..." line after last line.

		Issue: #324