Test: Auto load and store config data #339
The config.py will be used to get the data out of config.json.
The other files should not load json. It should be automatically
loaded and stored when creating instance of class Config

In class Config, initialization function loads json file. So
everytime an instance of Config class is made, json is automatically
loaded