Test: Load and merge two CSVs in pandas dataframe #339
The CSVs to be compared are to be mapped first from row to row
in each csv. It will be achieved using pandas merge method. Similar
rows will be merged together and then compared

Uploaded two csvs as pandas dataframes. Merged both the dataframes
on the basis of compoundId, Formula & Compound

Issue: #339