Ftr: Access groups from grouby objects using key #339
Similar groups will be compared. To do so, groups needs to be
divided on some features and then access these using key. Key is
combination of features used (or columns for this particular module)

Add function get_groups() to access list of groups from list of
pandas groupby objects using a key

Issue: #339