Fix: Add method definition to write samples #389
readSamplesXML will read sample name attribute of xml file
which name precedes with 's' as it's been updated while writing
to follow xml attribute naming which should start with a english
letter not any special character or number. If value of this attribute
is "Used", it will be added to samples attribute of PeakGroup
otherwise will be ignored.

Issue: #389