Fix: Remove corner case for single EIC during merging #334
When a single sample is selected, raw EIC is used as the merged EIC while for eicMerge() is called for multiple samples.
The additional smoothing step involved in creation of mergedEIC by eicMerge() helps avoid marking small spikes in the data as peaks.
Therefore a number of small peaks are seen in case of single sample that disappear when another sample is selected. Results are not consistent.

Issue: #334