Ftr: Update codebase for MassCutoff #137
	Here to introduce mDa along with ppm, float type is replaced
with custom type MassCutoff, and this class has a method
<massCutoffValue> which will take a mz value for which mass cutoff
has to be calculated, and based on value of string member <_massCutoffType>
which has to be set whenver mass cutoff type is changing (ppm to mDa
or vice-versa), it will return desired result. <_massCutoffType>
can hold "ppm" or "mDa" currently.

	Whole codebase is updated for this to incorporate this
type change. Where value of masscutoff <ppm or mDa> is required,
<getMassCutoff> method is used and to get result of actual value
( ppm*mz/10^6 or mDa),  <massCutoffValue>  is used.

Note: It's huge change in single commit. It was neccessary to maintain
code compilation and to avoid error. If some part were using float type
and another is using MassCutoff type, it could produce compilation
errors.It can also produce erroneous result if somewhere is used ppm
and somewhere mDa and at this commit, ElMaven is used. MassCutoff type
is prefered for feature adding purpose in future, which can be added in
almost 10 lines code.

Issue: #137