Fix: Use product m/z instead of base peak m/z #170
Base peak m/z is the m/z of the highest intensity ion in a scan.
It was stored as productMz and used in MS/MS peak detection.
This would give mostly correct results for MRM data but not for other MS/MS data types.

getEIC has been modified to go through every m/z of the correct precursor scan to find the appropriate product Mz.

Issue: #170