Ref: move getIsotopicMatrix to isotopeplot.cpp
	getIsotopicMatrix is only used by isotopeplot.cpp and it
should not be part of mainwindow.cpp, but this method is using
isotopeC13Correct. Now to remove dependencies from mainwindow.cpp,
isotopeC13Correct is updated as static member function which is not
just used by isotopeplot.cpp, isotopeplotwidget.cpp is also using it.
Making static solve this problem. But MavenParameters also has be
static