[Fix] Stop mass adjustment if mass is zero #493
Compound mass is set to zero for MS/MS fragments. It is used as a check for MS2 in multiple places.
AdjustMass was adding or removing proton mass even when mass is zero. This caused some of these checks to fail.
One prominent result: expectedMz showing "NA" in Peaks Table

Issues: #493