[Fix] Check for peaks before bookmarking a group
Using shift-drag integration, all peaks might be filtered out in some cases. Such a group should not be added to the list of groups.

In case of mzrolls, groups are added to the vector as we read the xml. Peaks are added to the groups at a later point.

The check for peaks has been moved to the bookmark function as shift-drag is the only way to get a group with no peaks.