Fix: Prevent pushing data to Polly when peaks table is empty
This commit will correctly stop the upload process if the peaks
table has no entries. ElMaven was sending compound DB file and
maven analysis settings to polly even if the peaks table was empty
and peak data was not sent. The user will also be notified of this
issue, through a message box.

Issue: #777