Fix sample widget slowdown with large peak table
Fixing an issue where having a large peak table would cause the
sample widget, when interacted with, to perform a heavy (blocking)
computation within the alignment visualisation widget(s).

Now this computation will only be performed if the alignment widget
is visible, and replotted whenever it is made visible using the
sidebar widget.

Issue: #723