Add private variable storing peak table type
TableDockWidget class is being used for three different purposes
and has not been subclassed for the same. Avoiding major changes,
this commit adds a private variable that stores the type of groups
that an instance of the widget represents. At the moment there can
be three different types of TableDockWidget: bookmark table, peak
table and scatterplot table.