[Ref] Move code to a new function for modularity
get_project_upload_url_commands is supposed to upload the given files and return the output and errors

It was taking a QByteArray and creating a json document just to get access to a string. This string is now the argument sent to the function.

Parsing the QByteArray and finding the string has been moved to a separate function

Documentation of some functions has been changed to the doxygen format