Always use absolute value of charge to adjust mass
When ionisation method is set to EI, a different formula is used
to calculate m/z value. This formula was not using the absolute
value of charge and negative m/z ratio is possible, which should
not happen and can potentially break analysis. This issue has been
corrected with this commit.

Issue: #841