[Fix] Upload links for Polly
Sample cohort validation check was failing as the code for deleting the files had been moved in a different branch. The validation function could not find the cohort file and the check failed.

There was also a rebase error that led to not using the getRedirectionUrl function resulting in firstview being the default url.