Fix the order of export for peak detailed format
The order of peaks exported (when using peak detailed format) was
different when run on different operating systems (tested on MacOS
and Ubuntu Linux). This commit will sort the order of peaks in a
group in the ascending order of the sample names for each peak,
thus ensuring that they are always exported in a certain format.