Fix missing labels in CSV exports after changed label settings
Once a peak detection operation has been completed and a table
generated, the isotope labels detected should be exported out as
they are shown in the peak table (which is also how they exist
internally as subgroups of each group). With this commit it is
ensured that even if isotope detection options are changed, all
isotope children detected are exported in the CSV report from a
peak table containing labelled data.

Issue: #882