[Fix] Resolve crash in merging tables
Merging the bookmarked groups to a peak table led to the app crashing if any peak table had been deleted during the session.
The merging code was working on the assumption that the table id and the peakList indices will remain in sync.
Deleting a peak table led to an index out of range error.
This has been resolved by searching for the correct Id