Allow saving as mzroll and mzrollDB from app menu
With this commit the user will now be able to save their session
in either mzroll or mzrollDB formats from the "File" app menu.
Additionally, the methods and button for saving mzroll files in
TableDockWidget class have been removed. The methods for mzroll
read/write will now be served by mzFileIO and ProjectDockWidget
classes in conjunction.