Change SQLite database format name to "emDB"
The .mzrollDB extensions are used by the original MAVEN project
and since the schema for El-MAVEN's database format differs in
certain aspects from the original one, .mzrollDB extensions can
not be opened in El-MAVEN and it would be wrong to assign the same
extension to databases with different schema formats. From this
point SQLite database outputs from El-MAVEN and inputs taken in
will be of the new El-MAVEN Database Format (.emDB).