Prompt user to save their session upon exit
The user will now be prompted to save their data when quitting a
session. There will be no prompt if there is already a ".emDB"
project open, which will be updated with the data when the
application is quit. This is also true for already open ".mzroll"
project being worked on.

As a side-effect of this patch, bookmarking autosave functionality
got broken. This should be fixed again in a subsequent patch.