Implement persistent autosave for sessions
Whenever El-MAVEN crashes, the user risks losing their progress
for that session. This calls for a need to have a somewhat
fail-safe autosaving mechanism that tries to minimize this loss
in the event of an abrupt application exit. A timestamped autsave
file will now be generated the first time a user bookmarks or marks
a peak group as good or bad. This file will be used to constantly
save any further curations done by the user in that session.
Finally a prompt will appear at the end of a session asking the
user if they want to save it as a named emDB file.