[GA] Add trigger for successful load of compound DB
The value for successful load is the MS level of the db.
When precursor and product m/z are present for a compound, all processing is done with the assumption that this is a MRM workflow.
The same rhetoric has been used to determine if the db is MS or MSMS.

This does not handles cases where the db has both MS and MSMS data as that is not a defined format.