Fix no save before new session load
Saving the data for a session as a project before opening a new
project was not happening even though the user was led to believe
it was through UI interactions. This commit fixes this issue by
not resetting autosave state before calling a method which relies
on its check for autosave state to do the correct thing (of saving
data for the exiting session) before resetting autosave and opening
a new session.