Fix loading emDB files crashing El-MAVEN
This patch replaces the existing signal connections to emDB load
progress dialog with a variant that takes in a receiver object
along with the slot function. Slots by default are not owned by
any thread, and if a receiver is not specified will be called from
the signalling thread instead of a receiving thread. This works
fine until there is a need for cross-thread communication using
signals and the sender and receiver objects live in two different
threads - as in our case.