[Fix] Show error message for obi-warp instead of exiting
There is a check in the 3rd party library, obiwarp, that calls for an exit if a check fails. This leads to the app closing abruptly during alignment. This is non-deterministic and tends to work if the process is re-started.
Until the reason for this process being non-deterministic is found, an error message has been put in place that will cancel the alignment if the check fails and ask the user to try again. This is just a temporary solution to prevent data loss.