Load compound databases in a separate thread
Compound databases, especially MSP files, can be large in size and
loading them is a potentially UI blocking operation. All compound
database loading will now happen in a separate loader thread, with
the exception of the first load that happens at application start,
in which case using a thread is not viable.

Issue: #965