Preserve all fragmentation records from MSP files
A molecule may give rise to fragmentation patterns depending on
the collision energy leading to splits. All of these patterns will
now be stored in compound database used by El-MAVEN, but the name
of each new fragmentation pattern for an existing compound will be
added with a numeric suffix signifying its load order into the
database. The "id" of each entry remains unchanged which is the
same as the original name of a PRM compound.

Issue: #965