Add versioning utilities to projectDB module
As the format of the database for storing information changes with
subsequent releases, there might be cases where the table structure
might need to be modified while keeping the data intact. A set of
functions and data structures have been added in the module for
allowing easy upgrades requiring minimal effort with each release
involving an upgrade.

Upgrades can be made by writing SQL scripts that transform the
data from one DB format to the next. For now, only backwards
compatibility is guaranteed by these utilities, because only
upgrades can be performed on database files from previous versions.
Older releases of the library will have no way of knowing how to
downgrade a newer version of the database to an older one that the
older version of the application would be able to use.

Issue: #969