Use versioning in ProjectDatabase class
All ProjectDatabase instances will now be able to perform checks
to evaluate whether they need a database upgrade and will
automatically attempt to do so.

Since database upgrades can potentially change the database to be
incompatible with older release in which it was first created, a
backup is also saved of the original DB before upgrading.

For development builds, an extra check is made to ensure that the
latest DB format available is used.

Issue: #969