Add methods of writing and reading user settings
With the new "user_settings" table added to the database, proper
methods are needed to insert, read and delete these values. These
methods have been added.

A new type has also been declared: `variant` which is a boost
variant capable of holding some primitive types and strings. It
has been added with the intention of enveloping different data
types possible for settings values. Conversion to and from this
type is simple and has been documented to some extent.

Issue: #969