Fix crash when changing mass resolution
Whenever isotope widget is open and peaks are selected, the widget
is supposed to be updated with isotopologue information of the
selected peak. However, the method that is called on peak selection
sets the current group after setting the sample (of the selected
peak) which causes a slot to be called which uses current group.
The slot may or may not be called before the group is set and this
was causing irregular crashes. This call order has been fixed with
this commit.

At the same time, the worker threads used for pulling isotopes have
been corrected with updating the correct MavenParameters object, to
further minimize the possibility of crashes.

Issue: #918