Add Polly QuantFit to EPI application interface
QuantFit is a service that allows quantification of metabolites as
well as that of labelled fragments of metabolites. With this
addition users will directly be able to send intensity data to
QuantFit through El-MAVEN. They will be provided with a link that
will take them to an unploader interface where they can provide
any required metadata for proceeding with calibration.

Issue: #984