Send email whenever users upload data to Polly
Users are presented with a button on EPI dialog to go to their
analysis or upload interface upon successful upload of data to
Polly. But they lose this link when they close the application. As
a measure of preserving analysis links, the user will now be sent
an email containing the URL from where they can proceed on to their
analysis (or upload) everytime.

Issue: #984