Allow multiple compounds with same ID
Until now compounds (other than PRM) were discarded if there was
a compound having the same ID already existed in the database. Now
every compound from CSV file (whose format is not malformed) will
be read in. Each copy of an existing compound will be added with
a number beside it for identification in ligand widget (and its
name changed internally as well).

Issue: #973