Delete old file when saving over existing project
Until now, overwriting an emDB file while saving a new one only
deleted the data inside the existing file and rewrite new data in
it. This did not cause any issues until we had project versioning.
Now the old file is opened as an older version of emDB file and the
versioning mechanism tries to make a backup of it. This has been
fixed. So, overwriting is done in the true sense -- old file is
deleted, then a new one is opened in which data is written.