[Fix] Copy current peakGroup in spectrawidget
Spectra widget used to store a pointer to the currently selected peakGroup from the peakTable.
On deleting the peakTable, the peakGroup is also deleted and therefore accessing it through the pointer make the app crash.

The peakGroup object will now we copied and stored as the currentGroup to prevent such crashes.