Close autosave DB projects before deletion
There is an incorrect check in place that was supposed to prevent
creation of a new SQLite project connection, if a save had been
requested for an already open project file. This was causing a new
database connection to be created for projects that already had an
open connection. The existing open connections were themselves
being lost, and not properly destroyed.

SQLite refuses from giving up write access on database files that
are in use or have unclosed connections. And this, in turn, was
not allowing El-MAVEN to be able to remove autosave project files
that were no longer needed (but had leaked connections hanging in
the memory), specifically on NTFS that uses a different scheme for
granting permissions.

With this patch, the check has been fixed and if an existing project
is already open, then another connection will not be made.
Additionally, all autosave files will have their connection closed
before attempting to delete them.

Issue: #1007