Borrow decoding and decompression logic from MAVEN
Decompression of zlib encoded data strings from sample files has
not been possible for a long time. This issue has already been
fixed in the upstream MAVEN project, by using a new base64 decoding
function that returns a string and then simply passing this string
to the decompression function in mzUtils (if decompression is
needed at all). This bypasses the original implementation where a
lot of raw buffer manipulation was being done before sending the
string to the decompression function. The same mechanism will not be
used in El-MAVEN.

A lot of unused methods have also been reomved from base64
namespace. In future if we need the other base64 encoding method,
we can borrow it from a standard library instead of maintaining one
for ourselves.

Additionally, no check for compression state was being made while
parsing of MzML chromatograms and spectrums. This has also been
fixed.

Issue: #47, #48