[Fix] Calculate MS2 stats for bookmarked groups
Number of MS2 events and MS2 scores were not being calculated if the user bookmarked a group by double-clicking the peak bubbles.

To calculate the scores as part of the peak grouping process, compound information is necessary.

Compound assignment was happening during the filtering process further ahead in the workflow, therefore compound info is now sent to the peak grouping method.

The compound assignment and MS2 stats calculation should be removed from the filtering process in the future.