Fix crash upon deleting all samples
Whenever the sample count for ProjectDockWidget reaches 0, the user
is shown a message that they should restart the application for a
fresh session. The message box itself is deleted when the user
presses any of the buttons on it, because of a Qt::WA_DeleteOnClose
flag set on the message box. But after closing the box (and its
immediate deletion), a check was being made to see whether the user
chose to restart and this check involved querying the (now deleted)
message box's clicked button. This was leading to a crash.