Optimize AsLS Smoothing algo for large vectors
AsLS has cubic time complexity and there is not much we can do
about it since the 3rd party solver is the bottleneck (cubic).
But, we can reduce the length of the intensity vector because, we
do not need to consider every point in the intensity signal. A
down-sampling method is used to shorten the intensity vector before
feeding it into the AsLS algorithm. The obtained baseline is then
up-sampled back to the size of original intensity vector.

Issue: #924