Add drag and drop support to ProjectDockWidget
Being able to arrange samples in a desired order has been requested
by some users. This patch enables them to do so, by dragging and
dropping samples within a sample set in Samples widget. However,
this change will also disable the sorting functionality that had
been available to users. This is not an intentional change but
rather a drawback of using QTreeWidget for viewing our samples.

Issue: #913