Add zlib wrapper source
Boost iostreams library provides convenient asbtractions over
zlib which handle the details elegantly. But since, zlib is an
external library, they do not ship this wrapper in their release
binaries. Therefore, we need to build the source file providing
zlib support as a part of our source tree. An additional macro
also needs to be defined.