Add a new class DownloadManager
This commit just adds a new class called DownloadManager.
It's meant to be a generic class, i.e It can be used by anyone
that requires something to be downloaded from internet.

It supports both asynchronous and synchronous requests.
Synchronous requests should  be used when absolute neccessary
and the file size is small