Disable EPI group subset options
The ability to export good groups should only be available if the
user has any good groups in their table. Similarly, Excluding bad
groups should not be an option if all groups are bad in a table.
This patch makes sure that these options are disabled in the
respective scenarios so that its not possible to send empty groups
file to Polly.