Change match compound behavior from hover to click
Moving the mouse around in the spectra widget sets the m/z value of
match compound widget, which is helpful but not the best UX, due to
the fact that its extremely difficult to preserve the set m/z value.
The same feature will now be available with mouse clicks instead of
hovering.

Issue: #888