Find best match when searching for cached slice
The method `MassSlices::sliceExists` checked for the presence of a
cached mzSlice object that contains the mz object, but only ever
returned a mzSlice that was barely different from an existing one.
This not only gave rise to duplicate groups from different slices
in the final table but also increased the time taken by the mass
slicing process. This has been fixed and now both m/z and rt values
are taken into account when finding the best matching slice in
cache.