Fix comparison of exponent formatted number strings
NumericTreeWidgetItem already has the ability to sort numeric
strings using their true values (and not their alphanumeric
representation). It uses QCollator for these comparisons. However,
QCollator is not able to compare numbers formatted in exponent
form (i.e., format having a significand and exponent). This patch
fixes this separately comparing the exponent and the significand if
a string is found to match a pattern of this format.

Isse: #481