Make all fragments visible for selected peak group
The default zoom level of the selected group in peak table has
been always decided by the current scan's m/z boundaries (min/max).
This is not always ideal when we are comparing consensus and
reference spectra. This change will now set the bounds of spectra
widget, based on the superset range of the observed m/z values and
the reference m/z values.

Issue: #1104