Combine OBI-Warp with segmented alignment
Segmented alignment used fractional distance based interpolation
within segments and has proved to work well for restoring retention
times from saved segments. But this same technique can also be
coupled with OBI-Warp, which tends to have slow runtimes when
operating on high-resolution (i.e., large number of scans) samples.

The idea is that we send only a sparse representation of the sample
(along rt axis) by sending scans separated by a constant interval.
OBI-Warp will do its job of aligning the samples based on the
information and the scans that were not sent for alignment can
instead have their retention times updated using segmented
alignment technique. This way we reduce the time taken for aligning
high-res datasets to a small window and allow runtime to scale only
with the number of samples for a given dataset.