Activate and populate EPI elements after failure
After an EPI request failure, the elements of the Polly interface
dialog remained inactive and the user will not be able to continue
with the same dialog instance (closing and reopening the dialog,
once the internet returns is the only option). This behaviour has
been fixed: UI elements are reactivated once the error message is
accepted.

Issue: #1109