Remove child from the children vector on delete
The class `PeakGroup` offers a method to delete a child but this
method was only clearing the space-intensive data structures for
this child and not removing it from the vector that stores
children of a peak group. This has been fixed. As a result, a
deleted child will no longer appear as a child of the peak group
in exported CSV (and possibly other places where it might have
been causing inconsistencies).

Issue: #261