Disable isotope detection for non-parent adducts
Detection of isotopes of non-parent adducts is not as
straightforward as simply adjusting masses of the parent ion (once
for adduct and once for isotopologue). Adducts which are generated
from knocking out ionic species containing any labelled ions might
have to be handled carefully. To prevent confusion, isotope
detection for non-parent adducts will remain disabled, for now.