Add logging for peak detector CLI tool
Logs are a very useful way of recording and persisting the output
from an ongoing process. It is also very easy to read them and
extract meaningful information for developers who might be
interested in debugging or querying the state of the process at
any point of time. Since, CLI is now being used by mutliple parties
for different kinds of integration, logging the output and response
from Polly APIs can help them at runtime to diagnose whether the
process completed successfully or whether it failed and its point
of failure in such a case.

This commit adds storage of general commandline output from a CLI
run.

Issue: #1143