Preserve window size over sessions
El-MAVEN has some in-built code written to read saved window size
from the last session. But the window size itself was being scaled
down by a factor of 0.7 and saved in application settings. This
led to El-MAVEN's window shrinking every new session. This scaling
has been removed and now El-MAVEN should spawn at the exact
position and size as when it was last (normally) closed.

Issue: #820