[Ftr] Warn user about sample load failure #676
There might be various reasons due to which some (or all) samples
can end up failing to load. Currently, there is no notification
on such events which can be confusing to the user. Ideally, the
exact reason for failure would help them understand what went
wrong. In future we can have a proper error propagation mechanism,
but for now we can at least inform users about the samples that
failed to load.