[Ftr] Read El-MAVEN's settings file in CLI #1163
El-MAVEN's setting (XML) file contains near-complete coverage of
the different options provided by the software. The peakdetector
CLI utility uses a subset of these options (if we exclude database
and sample paths), so it is possible to use the same settings file
as base config for the CLI. This makes the process of comparison
and reproduction of an analysis (on GUI) possible since the same
settings file can be used for both automated and GUI workflow.