[Fix] Pick MS2 events over individual peaks RT
During our PeakGroup-mzSlice refactoring (PR #686), accidentally
a change was introduced, which made calculation for MS2 group
(consensus) spectra get calculated for each sample over the entire
peak-group's RT coverage. Originally it was so that only those
MS2 scans for a sample were considered for which the fragmentation
RT lied within that sample's individiual peak. The post-refactor
method is incorrect (and was unintentional) which is being undone
with this commit.