[Fix] Incorrect peak count in results table #1142
Peak table shows the number of (non-blank) samples in which a peak
was detected as the "# peaks" value. If in presence of blank
samples, a peak-group has an above-threshold quality peak from the
blanks, then the good peak count (show as "# good") will include
them as well. The two metrics (total peak count and total good peak
count) should not be different in their consideration of blank
samples. This has been corrected. The displayed peak count will
now be shown as the sum of peak-counts from blank and non-blank
samples.