[Fix] Filter formula string for special characters #1168
Only a subset of alpha-numeric characters are read from formula
strings from any compound database. But internally the formula is
still saved as the same string that was provided with the database.
This, although it does not affect El-MAVEN much, can be problematic
for downstream analysis where valid chemical formula (as in KEGG DB
notation) may be expected. Therefore, with this commit, all formula
strings will be stripped as soon as they are assigned to a compound,
and exported to different formats/destinations as such.