[Fix] Missing mz-intensity pairs from scans
While iterating over an mz-intensity pair vector (obtained from
parsing mzXML spectra strings), we populate Scan::mz and
Scan::intensity vectors in the exact same indexes as recorded in
a scan. But some zero values could be present in-between two
non-zero pairs, and when the mz and intensity vectors are resized
to the non-zero count value, these zero pairs might end up staying
while non-zero pairs are the end of the vectors are cleaved.