[Ftr] Allow exporting sample metadata file template #1171
Editing metadata of samples within El-MAVEN, when the sample list
is large quickly becomes tedious. For this reason we have metadata
import facility. But generation of this metadata (CSV) file
separately, is itself tedious for large number of samples. With
this change users will be able to export a metadata template for
their current set of loaded samples, which they can easily edit
in a spreadsheet software and then use the import facility to get
the desired effect.