[Fix] Mark C12 PARENT peak-group as same as parent #537
Peak-groups when marked good or bad are exported with their labels,
except when they are isotopic parent groups. With this patch, along
with the parent group, the child group labelled as "C12 PARENT" is
also marked good/bad. This way when the table data is exported, the
curation will be visible for parent isotopic group at least. The
other isotopic groups will remain unmarked and left to the user's
discretion.