[ref] Resolved an issue in the doctest tests being performed
In earlier version of "jsonReports", the json file being generted was
tested against the saved json file. But the issue was that the file
generated is similar but the order of the group may differ from the
saved file.
This has been resolevd by comparing the groups corresponding groups in
the saved json file regardless of the order.