Visualize area under the curve for shift+drag #782
This commit adds a fade/highlight mode using either the "Shift"
key or the manual integration icon. During this "integration mode",
if the user drags their mouse across an RT range, the area under
the curve will be visually more clear. The integrated peak-group
will be added to the bookmark table (as it did previously).