[Fix] Re-enable running concurrent sessions #1215
The log files, that were introduced for some utilities in the last
release, were putting a restriction of being editable by only one
process (most prominently on Windows), which lead to a second
parallel session of El-MAVEN to throw a write exception as soon as
it started. To prevent this event, log files will now write to a
separate timestamped log, that share the same prefix for files
from a single session but are different for files of a different
El-MAVEN/peakdetector session.

Note: this fix also relies on the fact that, as soon as an
executable begins they create at least one log that is created
using the static timestamp - which is true for both El-MAVEN (which
creates a DownloadManager log) and peakdetector (which creates a
log for itself) - while starting up.