[Fix] Resolve compound suffixes over sessions #1003
Once a compound gets saved with its suffixed name in an emDB, its
original name is lost and when the emDB is re-loaded, it may not
get compared against it's existing identical compound whose name
is still the original one. This has been fixed and compounds
stored in emDB will be restored near-identically.