[UI] Add new mode for upload to Polly project
Users may want to use services on Polly, that they have licenses
for, but is not part of the three existing integrations (FirstView,
PollyPhi and QuantFit). Currently, they will have to export their
files and upload to Polly as two separate steps. Instead we are
adding a "Send to Project" feature, where they can simply upload
their data to Polly without having a workflow associated to it.