[fix] Fixed the issue of missing peak detection for isotopologues
In case of C13 isotopic detection, some EIC were represented on the
graph which do not have the corresponding peaks. That gave the user
the view of missing peaks.
This has been corrected by adding a check that only those sample's  EIC
will be represented if there is a corresponding peak for it.