[ref] Refactored exporting of Eic's as pdf to make it work a thread
Earlier exporting EIC's as pdf took a long time and the user would feel
that ElMaven is not responding. Due to the rendering of pdf, the user
screen was at hault. This was not a good user experience.
Now, exporting pdf is done in a separate thread that does not hault the
screen and provides notifications to the user when pdf has been exported.